backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pdfminer.six
python-docx
spacy
reportlab
google-generativeai
python-dotenv
requests
beautifulsoup4
